Isochrones : code refactoring
Create a new class JourneyCommon with duplicate code in Journeys and
GraphhicalIsochrones.
Add data_freshness param for GraphicalIsochrones.
