+++ source/jormungandr/jormungandr/scenarios/simple.py
@@ -39,6 +39,14 @@ from jormungandr.scenarios.utils import build_pagination
 from jormungandr.scenarios.utils import updated_common_journey_request_with_default
 
 
+def get_pb_data_freshness(request):
+    if request['data_freshness'] == 'realtime':
+        return type_pb2.REALTIME
+    elif request['data_freshness'] == 'adapted_schedule':
+        return type_pb2.ADAPTED_SCHEDULE
+    else:
+        return type_pb2.BASE_SCHEDULE
+
 class Scenario(object):
     """
     the most basic scenario, it's so simple it don't implements journeys!
@@ -198,15 +206,15 @@ class Scenario(object):
             origin = journey_req.origin.add()
             origin.place = request["origin"]
             origin.access_duration = 0
-            journey_req.clockwise = True
         if "destination" in request and request["destination"]:
             destination = journey_req.destination.add()
             destination.place = request["destination"]
             destination.access_duration = 0
-            journey_req.clockwise = False
         request["datetime"] = [request["datetime"]]
         for dte in request["datetime"]:
             journey_req.datetimes.append(dte)
+        journey_req.clockwise = request["clockwise"]
+        journey_req.realtime_level = get_pb_data_freshness(request)
         updated_common_journey_request_with_default(request, instance)
         sn_params = journey_req.streetnetwork_params
         sn_params.max_walking_duration_to_pt = request["max_walking_duration_to_pt"]
