+++ source/jormungandr/tests/check_utils.py
@@ -990,6 +990,8 @@ r_coord = "0.00188646;0.00071865"  # coordinate of R in the dataset
 #default journey query used in various test
 journey_basic_query = "journeys?from={from_coord}&to={to_coord}&datetime={datetime}"\
     .format(from_coord=s_coord, to_coord=r_coord, datetime="20120614T080000")
+isochrone_basic_query = "isochrones?from={from_coord}&datetime={datetime}&max_duration={max_duration}"\
+    .format(from_coord=s_coord, datetime="20120614T080000", max_duration="3600")
 
 def get_all_disruptions(elem, response):
     """
